requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
asttokens=2.1.0=pypi_0
astunparse=1.6.3=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=pypi_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.2=pypi_0
flatbuffers=22.10.26=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
google-pasta=0.2.0=pypi_0
h5py=3.7.0=pypi_0
idna=3.4=pypi_0
imageio=2.22.4=pypi_0
ipykernel=6.17.1=pypi_0
ipython=8.6.0=pypi_0
jedi=0.18.1=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=7.4.5=pypi_0
jupyter-core=5.0.0=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.39=hc81fddc_0
libclang=14.0.6=pypi_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgomp=12.2.0=h65d4601_19
libnsl=2.0.0=h7f98852_0
libsqlite=3.39.4=h753d276_0
libuuid=2.32.1=h7f98852_1000
libzlib=1.2.13=h166bdaf_4
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pypi_0
networkx=2.8.8=pypi_0
numpy=1.23.4=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
openssl=3.0.7=h166bdaf_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.5.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=pypi_0
pip=22.3.1=pyhd8ed1ab_0
platformdirs=2.5.3=pypi_0
plotly=5.11.0=pypi_0
prompt-toolkit=3.0.32=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pygments=2.13.0=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.2=pypi_0
python=3.9.13=h2660328_0_cpython
python-dateutil=2.8.2=pypi_0
pytz=2022.6=pypi_0
pywavelets=1.4.1=pypi_0
pyzmq=24.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
setuptools=65.5.1=pyhd8ed1ab_0
six=1.16.0=pypi_0
sklearn=0.0.post1=pypi_0
sqlite=3.39.4=h4ff8645_0
stack-data=0.6.0=pypi_0
tenacity=8.1.0=pypi_0
tensorflow=2.10.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
termcolor=2.1.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
tk=8.6.12=h27826a3_0
tornado=6.2=pypi_0
traitlets=5.5.0=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022f=h191b570_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
xz=5.2.6=h166bdaf_0
